fix(store): Fix store max offset of the specified stream/range (#396)
* WIP: Prepare to change WriteTask


* Record handle extension now is variant of hash-of-tag if it is a single entry record, or number of nested entries if composite


* fix #391 Store#max_offset of a given stream/range in data-node


* Use u32 to represent nested record entries number


* format


---------

+++ components/model/src/batch.rs
@@ -3,7 +3,7 @@ pub trait Batch {
     fn offset(&self) -> u64;
 
     /// Quantity of the internal nested records
-    fn len(&self) -> usize;
+    fn len(&self) -> u32;
 
     /// Indicate whether the batch is empty
     fn is_empty(&self) -> bool {
