+++ data-node/src/stream_manager/window.rs
@@ -13,7 +13,7 @@ pub(crate) struct Window<R> {
     requests: BinaryHeap<R>,
 
     /// Submitted request offset to batch size.
-    submitted: HashMap<u64, usize>,
+    submitted: HashMap<u64, u32>,
 }
 
 impl<R> Window<R>
@@ -88,7 +88,7 @@ mod tests {
     #[derive(Debug)]
     struct Foo {
         offset: u64,
-        len: usize,
+        len: u32,
     }
 
     impl Batch for Foo {
@@ -96,7 +96,7 @@ mod tests {
             self.offset
         }
 
-        fn len(&self) -> usize {
+        fn len(&self) -> u32 {
             self.len
         }
     }
@@ -175,11 +175,11 @@ mod tests {
 
         // After fast-forward, an inflight batch entry is inserted.
         assert!(window.fast_forward(&foo1));
-        // When commit, the offset should be amended if there is a corresponding inflight batch entry. 
+        // When commit, the offset should be amended if there is a corresponding inflight batch entry.
         // After the commit , the entry will be removed.
         assert_eq!(2, window.commit(0));
         assert_eq!(0, window.commit(0));
-        
+
         // If there is no inflight batch entry, the commit offset should not be amended.
         assert_eq!(2, window.commit(2));
